fs: Workaround the compiler's bad optimization
When compiling the kernel with  the compiler
"aarch64-linux-android-gcc (GCC) 4.9.x-google 20140827 (prerelease)"
The compiler seems to be optimizing a value out in the do_sync_write
function and is incorrectly passing the pointer instead of the
dereferenced value.
Force the compiler to use the right value by passing the dereferenced
pointer instead of another variable that is assigned with the dereferenced
value.

Change-Id: I60505ffe39393f6323dcc7c6f912c74ea6aca6cd
Signed-off-by: Nikhilesh Reddy <reddyn@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>
Signed-off-by: AudioGod <audiogod@sonic-developers.com>
Signed-off-by: Haikal Izzuddin <webhaikal@gmail.com>